Add CircleCI config for build and release
We need to automate the release process and also utilize Github Releases
as a simple distribution medium so people can download packages directly
from Github.

We use CircleCI because it works great with containers. The build stage
uses nginx-vts-builder public image and invokes build.sh script.

The release stage uses artifacts from build stage via workspace feature
and then uploads the packages to the Github Releases using ghr tool.

build and release jobs are run only on tags because the idea is to
create tag with nginx and vts versions like 1.15.7-0.1.18 and have a
package with these versions on Github Releases.